добавил третье задание

+++ src/main/java/Employee.java
@@ -1,5 +1,5 @@
 public class Employee {
-    public long id;
+    public int id;
     public String firstName;
     public String lastName;
     public String country;
@@ -8,7 +8,7 @@ public class Employee {
     public Employee() {
     }
 
-    public Employee(long id, String firstName, String lastName, String country, int age) {
+    public Employee(int id, String firstName, String lastName, String country, int age) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
